+++ properties/Connection/AnInvalidLazyQueryMustThrow.php
@@ -21,7 +21,7 @@ final class AnInvalidLazyQueryMustThrow implements Property
 {
     public static function any(): Set
     {
-        return Set\Elements::of(new self);
+        return Set::of(new self);
     }
 
     public function applicableTo(object $connection): bool