+++ properties/Connection/CreateTable.php
@@ -33,11 +33,11 @@ final class CreateTable implements Property
 
     public static function any(): Set
     {
-        return Set\Composite::immutable(
+        return Set::compose(
             static fn(...$args) => new self(...$args),
             Name::any(),
             Column::list(),
-        );
+        )->toSet();
     }
 
     public function applicableTo(object $connection): bool