+++ properties/Connection/SelectOrder.php
@@ -47,13 +47,15 @@ final class SelectOrder implements Property
 
     public static function any(): Set
     {
-        return Set\Composite::immutable(
+        return Set::compose(
             static fn(...$args) => new self(...$args),
-            Set\Uuid::any(),
-            Set\Uuid::any(),
-            Set\Strings::madeOf(Set\Chars::ascii())->between(0, 254),
-            Set\Integers::any(),
-        );
+            Set::uuid(),
+            Set::uuid(),
+            Set::strings()
+                ->madeOf(Set::strings()->chars()->ascii())
+                ->between(0, 254),
+            Set::integers(),
+        )->toSet();
     }
 
     public function applicableTo(object $connection): bool