+++ properties/Connection/SelectWhereInQuery.php
@@ -49,17 +49,23 @@ final class SelectWhereInQuery implements Property
 
     public static function any(): Set
     {
-        return Set\Composite::immutable(
+        return Set::compose(
             static fn(...$args) => new self(...$args),
-            Set\Uuid::any(),
-            Set\Uuid::any(),
-            Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
-            Set\Integers::any(),
+            Set::uuid(),
+            Set::uuid(),
+            Set::strings()
+                ->madeOf(Set::strings()->chars()->ascii())
+                ->between(0, 255),
+            Set::integers(),
             Set\MutuallyExclusive::of(
-                Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
-                Set\Strings::madeOf(Set\Chars::ascii())->between(0, 255),
+                Set::strings()
+                    ->madeOf(Set::strings()->chars()->ascii())
+                    ->between(0, 255),
+                Set::strings()
+                    ->madeOf(Set::strings()->chars()->ascii())
+                    ->between(0, 255),
             ),
-        );
+        )->toSet();
     }
 
     public function applicableTo(object $connection): bool